+++ src/abandon.rs
@@ -4,21 +4,24 @@ use lber::structures::{Tag, Integer};
 use lber::common::TagClass;
 
 use futures::{future, Future};
-use tokio_proto::multiplex::RequestId;
 use tokio_service::Service;
 
 use ldap::{bundle, Ldap, LdapOp, next_req_controls};
+use protocol::LdapRequestId;
 
 impl Ldap {
-    /// See [`LdapConn::abandon()`](struct.LdapConn.html#method.abandon).
-    pub fn abandon(&self, id: RequestId) ->
+    /// Abandon the request identified by `msgid`. Since this requires knowing the
+    /// `msgid` of an operation, it currently works only for synchronous streaming
+    /// searches and asynchronous searches in general, invoked via methods on
+    /// structures representing those searches. See [`EntryStream::abandon()`]
+    /// (struct.EntryStream.html#method.abandon) and [`SearchStream::get_abandon_channel()`]
+    /// (struct.SearchStream.html#method.get_abandon_channel).
+    pub fn abandon(&self, msgid: LdapRequestId) ->
             Box<Future<Item=(), Error=io::Error>> {
         let bundle = bundle(self);
-        let msgid = match bundle.borrow().search_helpers.get(&id) {
-            Some(helper) => helper.msgid,
-            None => return Box::new(future::err(io::Error::new(io::ErrorKind::Other, format!("id {} not a search operation", id)))),
-        };
-        bundle.borrow_mut().abandoned.insert(id);
+        if !bundle.borrow().id_map.contains_key(&msgid) {
+            return Box::new(future::err(io::Error::new(io::ErrorKind::Other, format!("msgid {} not an active operation", msgid))));
+        }
         let req = Tag::Integer(Integer {
             id: 16,
             class: TagClass::Application,